requirements.txt
# LLM API Client
openai>=1.12.0

# PDF Processing
pdfplumber>=0.10.0

# Document Generation
python-docx>=1.1.0

# Configuration
PyYAML>=6.0.1

# Progress Bar
tqdm>=4.66.0
